+++ docs/make.jl
@@ -2,9 +2,11 @@ cd(@__DIR__)
 using Attractors
 using Attractors.DynamicalSystemsBase
 using Attractors.StateSpaceSets
+using PredefinedDynamicalSystems
 
 pages = [
     "index.md",
+    "dynsysref.md",
     "attractors.md",
     "basins.md",
     "continuation.md",